pyproject.toml
[project]
requires-python = "~=3.13"
dependencies = [
    "django~=5.1",
    "shortuuid~=1.0",
    "django-countries~=7.6",
    "django-docopt-command~=1.1",
    "django-admin-sortable2~=2.2",
    "django-convenient-formsets~=2.0",
    "django-widget-tweaks~=1.5",
    "djangorestframework~=3.15",
    "django-markdownify~=0.9",
    "psycopg[binary]~=3.2",
    "django-anymail~=12.0",
    "django-vite~=3.0",
    "pillow~=10.4",
    "django-htmx~=1.19",
    "django-decorator-include~=3.0",
    "django-easy-audit~=1.3",
    "setuptools~=75.1",
]

[tool.pdm.dev-dependencies]
dev = [
    "fabric~=3.2",
]

[tool.pdm]
distribution = false
